requirements.txt
numpy==2.1.2
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
pyqtgraph==0.13.7
pyserial==3.5
